rename pkg/prometheus to pkg/prom
