+++ pkg/config/config.go
@@ -4,7 +4,7 @@ import (
 	"flag"
 	"io/ioutil"
 
-	"github.com/grafana/agent/pkg/prometheus"
+	"github.com/grafana/agent/pkg/prom"
 	"github.com/pkg/errors"
 	"github.com/weaveworks/common/server"
 	"gopkg.in/yaml.v2"
@@ -12,8 +12,8 @@ import (
 
 // Config contains underlying configurations for the agent
 type Config struct {
-	Server     server.Config     `yaml:"server"`
-	Prometheus prometheus.Config `yaml:"prometheus,omitempty"`
+	Server     server.Config `yaml:"server"`
+	Prometheus prom.Config   `yaml:"prometheus,omitempty"`
 }
 
 // ApplyDefaults sets default values in the config
