+++ pkg/prom/ha/configapi/types.go
