+++ pkg/prom/ha/lazy_transferer.go
