+++ pkg/prom/instance/instance.go
@@ -16,7 +16,7 @@ import (
 
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
-	"github.com/grafana/agent/pkg/prometheus/wal"
+	"github.com/grafana/agent/pkg/prom/wal"
 	"github.com/oklog/run"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/common/model"
