+++ pkg/prom/wal/util.go
